backend/requirements.txt
fastapi
uvicorn[standard]
python-multipart
pillow
colorthief
opencv-python
scikit-image
numpy